requirements.txt
############################################################
# Frontdesk Dental Clinic Backend - Requirements & Setup
# ----------------------------------------------------------
# Step 1: Make sure you are in a virtual environment
# Windows:
#   python -m venv venv
#   .\venv\Scripts\activate
# macOS / Linux:
#   python3 -m venv venv
#   source venv/bin/activate
#
# Step 2: Upgrade pip, setuptools, wheel
#   python -m pip install --upgrade pip setuptools wheel
#
# Step 3: Create a .env file in the project root with your
#   Django secret key, Supabase credentials, and CORS origins.
#
# Step 4: Install all dependencies with:
#   pip install -r requirements.txt
#
# Step 5: Run database migrations:
#   python manage.py migrate
#
# Step 6: (Optional) Create a superuser:
#   python manage.py createsuperuser
#
# Step 7: Run development server:
#   python manage.py runserver
############################################################

# Django web framework
Django==5.2.6

# ASGI / WSGI utilities
asgiref==3.9.1

# REST API framework
djangorestframework==3.16.1

# PostgreSQL database driver (Windows-compatible)
psycopg2-binary==2.9.10

# WSGI server for production
gunicorn==23.0.0

# CORS handling for frontend-backend communication
django-cors-headers==4.4.0

# Environment variable management
python-decouple==3.8

# Utility for parsing SQL (used by Django)
sqlparse==0.5.3

# Packaging utilities
packaging==25.0

############################################################
# 🔹 New dependencies for Supabase Auth (JWT verification)
############################################################

# JWT decoding & validation
PyJWT[crypto]==2.9.0

# Fetch Supabase JWKS (JSON Web Key Set) from URL
python-jose==3.3.0
cryptography==43.0.3

# HTTP client for fetching JWKS if needed
requests==2.32.3

# (Optional) Supabase Python client – if you want to query Supabase directly
supabase==2.10.0
